pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "network-monitoring"
version = "1.0.0"
description = "AI-powered network performance anomaly detection and automated ticketing system"
authors = [
    {name = "Farid Jabbari Maleki", email = "performance-tools@rightel.ir"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Telecommunications Industry",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: System :: Monitoring",
    "Topic :: Communications",
]
dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "prophet>=1.1.4",
    "scikit-learn>=1.3.0",
    "cx_Oracle>=8.3.0",
    "sqlalchemy>=2.0.0",
    "requests>=2.31.0",
    "urllib3>=2.0.0",
    "beautifulsoup4>=4.12.0",
    "Jinja2>=3.1.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "plotly>=5.15.0",
    "PyYAML>=6.0",
    "python-dotenv>=1.0.0",
    "Pillow>=10.0.0",
    "pydantic>=2.0.0",
    "typing-extensions>=4.7.0",
    "click>=8.1.0",
    "cryptography>=41.0.0",
    "python-dateutil>=2.8.0",
    "pytz>=2023.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]
async = [
    "asyncio-mqtt>=0.13.0",
    "aiohttp>=3.8.0",
]
monitoring = [
    "prometheus-client>=0.17.0",
]

[project.scripts]
network-monitor = "src.scripts.run_monitoring:main"

[project.urls]
"Homepage" = "https://github.com/your-org/network-monitoring"
"Bug Reports" = "https://github.com/your-org/network-monitoring/issues"
"Source" = "https://github.com/your-org/network-monitoring"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.setuptools.package-data]
"src" = ["templates/**/*", "config/**/*"]

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = [
    "tests",
]
python_files = [
    "test_*.py",
    "*_test.py",
]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test_*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
] 